+++ src/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v1.0.0
+
+Added four new commands:
+
+- **Next shade** - to select the next (darker) shade of the same colour
+- **Previous shade** - to select the previous (lighter) shade of the same colour
+- **Next colour** - to select the next text colour from the colour library
+- **Previous colour** - to select the previous text colour from the colour library
+
+### Limitations in v1.0.0
+
+Currently, the colours come from a library named "Colours", because that's what mine is called.
+
 ## v0.4.1 - 2024/08/06
 
 Fixes an incorrect import path for font size changes that occurred during file restructure.
@@ -11,7 +24,7 @@ Added two new commands:
 - **Increase font weight** - to select the next font weight from a pre-defined list
 - **Decrease font weight** - to select the previous font weight from a pre-defined list
 
-### Limitations
+### Limitations in v0.4.0
 
 Currently, the list of weights is pre-defined and includes Regular, Medium, Semibold and Bold.
 
@@ -35,7 +48,7 @@ In both cases:
 - If no text style is already selected, the plugin looks for a style with a font size of 16 to apply
 - If the current text style is already at the end of the list, based on the current direction, no changes are made
 
-### Limitations
+### Limitations in v0.1.0
 
 - The plugin currently expects the library containing the fonts to be named "Fonts", because that's what mine is called.
 - The plugin only looks for size 16 as the default.